+++ Command/ModifyCloudCommand.php
@@ -11,23 +11,125 @@
 
 namespace Xabbuh\PandaBundle\Command;
 
+use Http\Client\HttpClient;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Xabbuh\PandaClient\Api\AccountManagerInterface;
+use Xabbuh\PandaClient\Api\Cloud;
+use Xabbuh\PandaClient\Api\HttplugClient;
+use Xabbuh\PandaClient\Transformer\TransformerRegistry;
+
+if (class_exists(ContainerAwareCommand::class)) {
+    /**
+     * Modify a panda cloud via command-line.
+     *
+     * @author Christian Flothmann <christian.flothmann@xabbuh.de>
+     *
+     * @final since 1.5
+     */
+    class ModifyCloudCommand extends ContainerAwareCommand
+    {
+        use ModifyCloudCommandTrait;
+
+        protected static $defaultName = 'panda:cloud:modify';
+
+        private $accountManager;
+        private $transformerRegistry;
+        private $httpClient;
+
+        public function __construct(AccountManagerInterface $accountManager = null, TransformerRegistry $transformerRegistry = null, HttpClient $httpClient = null)
+        {
+            if (null === $accountManager) {
+                @trigger_error(sprintf('Not injecting an %s instance into the constructor of the %s class is deprecated since PandaBundle 1.5.', AccountManagerInterface::class, static::class), E_USER_DEPRECATED);
+            }
+
+            if (null === $transformerRegistry) {
+                @trigger_error(sprintf('Not injecting an %s instance into the constructor of the %s class is deprecated since PandaBundle 1.5.', TransformerRegistry::class, static::class), E_USER_DEPRECATED);
+            }
+
+            parent::__construct();
+
+            $this->accountManager = $accountManager;
+            $this->transformerRegistry = $transformerRegistry;
+            $this->httpClient = $httpClient;
+        }
+
+        public function setContainer(ContainerInterface $container = null)
+        {
+            @trigger_error(sprintf('The %s() method is deprecated since PandaBundle 1.5.', __METHOD__), E_USER_DEPRECATED);
+
+            parent::setContainer($container);
+        }
+
+        public function getContainer()
+        {
+            @trigger_error(sprintf('The %s() method is deprecated since PandaBundle 1.5.', __METHOD__), E_USER_DEPRECATED);
+
+            return parent::getContainer();
+        }
+    }
+} else {
+    /**
+     * Modify a panda cloud via command-line.
+     *
+     * @author Christian Flothmann <christian.flothmann@xabbuh.de>
+     *
+     * @final since 1.5
+     */
+    class ModifyCloudCommand extends Command
+    {
+        use ModifyCloudCommandTrait;
+
+        protected static $defaultName = 'panda:cloud:modify';
+
+        private $accountManager;
+        private $transformerRegistry;
+        private $httpClient;
+        private $container;
+
+        public function __construct(AccountManagerInterface $accountManager = null, TransformerRegistry $transformerRegistry = null, HttpClient $httpClient = null)
+        {
+            if (null === $accountManager) {
+                @trigger_error(sprintf('Not injecting an %s instance into the constructor of the %s class is deprecated since PandaBundle 1.5.', AccountManagerInterface::class, static::class), E_USER_DEPRECATED);
+            }
+
+            if (null === $transformerRegistry) {
+                @trigger_error(sprintf('Not injecting an %s instance into the constructor of the %s class is deprecated since PandaBundle 1.5.', TransformerRegistry::class, static::class), E_USER_DEPRECATED);
+            }
+
+            parent::__construct();
+
+            $this->accountManager = $accountManager;
+            $this->transformerRegistry = $transformerRegistry;
+            $this->httpClient = $httpClient;
+        }
+
+        public function setContainer(ContainerInterface $container = null)
+        {
+            @trigger_error(sprintf('The %s() method is deprecated since PandaBundle 1.5.', __METHOD__), E_USER_DEPRECATED);
+
+            $this->container = $container;
+        }
+
+        public function getContainer()
+        {
+            @trigger_error(sprintf('The %s() method is deprecated since PandaBundle 1.5.', __METHOD__), E_USER_DEPRECATED);
+
+            return $this->container;
+        }
+    }
+}
 
 /**
- * Modify a panda cloud via command-line.
- *
- * @author Christian Flothmann <christian.flothmann@xabbuh.de>
- *
- * @final since 1.5
+ * @internal
  */
-class ModifyCloudCommand extends ContainerAwareCommand
+trait ModifyCloudCommandTrait
 {
-    protected static $defaultName = 'panda:cloud:modify';
-
     /**
      * {@inheritDoc}
      */
@@ -77,15 +179,33 @@ class ModifyCloudCommand extends ContainerAwareCommand
 
         // if anything has changed
         if (count($data) > 0) {
-            // push these changes to the panda service
-            $cloudFactory = $this->getContainer()->get('xabbuh_panda.cloud_factory');
-            $cloud = $cloudFactory->get(
-                $input->getArgument('cloud-id'),
-                $input->getOption('account')
-            );
+            $cloud = $this->getCloud($input->getArgument('cloud-id'), $input->getOption('account'));
             $cloud->setCloud($data, $input->getArgument('cloud-id'));
         }
 
         return 0;
+    }
+
+    private function getCloud(string $cloudId, ?string $accountKey): Cloud
+    {
+        if (null === $this->accountManager) {
+            $this->accountManager = $this->getContainer()->get('xabbuh_panda.account_manager');
+        }
+
+        if (null === $this->transformerRegistry) {
+            $this->transformerRegistry = $this->getContainer()->get('xabbuh_panda.transformer');
+        }
+
+        $account = $this->accountManager->getAccount($accountKey);
+
+        $httpClient = new HttplugClient($this->httpClient);
+        $httpClient->setCloudId($cloudId);
+        $httpClient->setAccount($account);
+
+        $cloud = new Cloud();
+        $cloud->setHttpClient($httpClient);
+        $cloud->setTransformers($this->transformerRegistry);
+
+        return $cloud;
     }
 }